server/requirements.txt
# Optional (development)
pytest
pytest-django
ipython

# Core
asgiref==3.11.0
Django==5.2.9
setuptools==65.5.0
sqlparse==0.5.4
tzdata==2025.2
djangorestframework
django-cors-headers
python-dotenv
django-filter==23.3
django-storages==1.14.2
django-environ==0.11.2

# Authentication
djangorestframework-simplejwt
social-auth-app-django==5.2.0
PyJWT==2.8.0

# Database
mysqlclient
psycopg2-binary==2.9.9
django-jsonfield==1.4.1

# File handling
Pillow==10.1.0
boto3==1.34.0

# Payments
stripe==7.0.0

# Email / Background tasks
django-sendgrid-v5==0.9.0
celery==5.3.1

# Development tools
drf-yasg==1.21.7
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.3.0
redis==5.0.1
